// packages/constants/src/file/type.ts
export const enum FileSuffixEnum {
  /** 3GP视频文件 */
  '3gp' = '3gp',
  /** 7-Zip压缩文件 */
  '7z' = '7z',
  /** AAC音频文件 */
  aac = 'aac',
  /** AC3音频文件 */
  ac3 = 'ac3',
  /** Adobe Illustrator文件 */
  ai = 'ai',
  /** AIFF音频文件 */
  aiff = 'aiff',
  /** Apple Lossless音频文件 */
  alac = 'alac',
  /** AMR音频文件 */
  amr = 'amr',
  /** ARC压缩文件 */
  arc = 'arc',
  /** ARJ压缩文件 */
  arj = 'arj',
  /** Advanced Systems Format视频文件 */
  asf = 'asf',
  /** AU音频文件 */
  au = 'au',
  /** AVIF图像文件 */
  avif = 'avif',
  /** AVI视频文件 */
  avi = 'avi',
  /** BMP图像文件 */
  bmp = 'bmp',
  /** Brotli压缩文件 */
  br = 'br',
  /** Bzip压缩文件 */
  bz = 'bz',
  /** Bzip2压缩文件 */
  bz2 = 'bz2',
  /** CAB压缩文件 */
  cab = 'cab',
  /** CD音频文件 */
  cda = 'cda',
  /** CSV表格文件 */
  csv = 'csv',
  /** DivX视频文件 */
  divx = 'divx',
  /** Word文档文件 */
  doc = 'doc',
  /** Word文档文件(XML) */
  docx = 'docx',
  /** DV视频文件 */
  dv = 'dv',
  /** DTS音频文件 */
  dts = 'dts',
  /** Enterprise Archive文件 */
  ear = 'ear',
  /** Encapsulated PostScript文件 */
  eps = 'eps',
  /** Flash视频文件 */
  f4v = 'f4v',
  /** FLAC音频文件 */
  flac = 'flac',
  /** Flash视频文件 */
  flv = 'flv',
  /** GIF图像文件 */
  gif = 'gif',
  /** GSM音频文件 */
  gsm = 'gsm',
  /** Gzip压缩文件 */
  gzip = 'gzip',
  /** H.264视频编码文件 */
  h264 = 'h264',
  /** H.265视频编码文件 */
  h265 = 'h265',
  /** HEIC图像文件 */
  heic = 'heic',
  /** HEIF图像文件 */
  heif = 'heif',
  /** HEVC视频编码文件 */
  hevc = 'hevc',
  /** ICO图标文件 */
  ico = 'ico',
  /** ISO镜像文件 */
  iso = 'iso',
  /** JFIF图像文件 */
  jfif = 'jfif',
  /** Java Archive文件 */
  jar = 'jar',
  /** JPEG图像文件 */
  jpeg = 'jpeg',
  /** JPG图像文件 */
  jpg = 'jpg',
  /** LHA压缩文件 */
  lha = 'lha',
  /** LZH压缩文件 */
  lzh = 'lzh',
  /** LZ压缩文件 */
  lz = 'lz',
  /** LZMA压缩文件 */
  lzma = 'lzma',
  /** LZO压缩文件 */
  lzo = 'lzo',
  /** MPEG-2传输流文件 */
  m2ts = 'm2ts',
  /** M4A音频文件 */
  m4a = 'm4a',
  /** M4V视频文件 */
  m4v = 'm4v',
  /** MIDI音频文件 */
  mid = 'mid',
  /** MIDI音频文件 */
  midi = 'midi',
  /** Matroska视频文件 */
  mkv = 'mkv',
  /** MOD视频文件 */
  mod = 'mod',
  /** QuickTime视频文件 */
  mov = 'mov',
  /** MP3音频文件 */
  mp3 = 'mp3',
  /** MP4视频文件 */
  mp4 = 'mp4',
  /** Musepack音频文件 */
  mpc = 'mpc',
  /** MPEG视频文件 */
  mpg = 'mpg',
  /** AVCHD视频文件 */
  mts = 'mts',
  /** Matroska音频文件 */
  mka = 'mka',
  /** Ogg视频文件 */
  ogv = 'ogv',
  /** Opus音频文件 */
  opus = 'opus',
  /** PAK压缩文件 */
  pak = 'pak',
  /** PDF文档文件 */
  pdf = 'pdf',
  /** PCM音频文件 */
  pcm = 'pcm',
  /** PNG图像文件 */
  png = 'png',
  /** PowerPoint演示文件 */
  ppt = 'ppt',
  /** PowerPoint演示文件(XML) */
  pptx = 'pptx',
  /** Photoshop文件 */
  psd = 'psd',
  /** RAR压缩文件 */
  rar = 'rar',
  /** Real Audio文件 */
  ra = 'ra',
  /** RAW图像文件 */
  raw = 'raw',
  /** Real Media文件 */
  rm = 'rm',
  /** Real Media Variable Bitrate文件 */
  rmvb = 'rmvb',
  /** RZIP压缩文件 */
  rz = 'rz',
  /** SEA压缩文件 */
  sea = 'sea',
  /** StuffIt压缩文件 */
  sit = 'sit',
  /** StuffIt X压缩文件 */
  sitx = 'sitx',
  /** Speex音频文件 */
  spx = 'spx',
  /** SVG矢量图像文件 */
  svg = 'svg',
  /** Shockwave Flash文件 */
  swf = 'swf',
  /** TAR压缩文件 */
  tar = 'tar',
  /** Bzip2压缩的TAR文件 */
  tbz2 = 'tbz2',
  /** TIFF图像文件 */
  tiff = 'tiff',
  /** Gzip压缩的TAR文件 */
  tgz = 'tgz',
  /** TOD视频文件 */
  tod = 'tod',
  /** MPEG传输流文件 */
  ts = 'ts',
  /** True Audio文件 */
  tta = 'tta',
  /** UUEncoded文件 */
  uue = 'uue',
  /** VOC音频文件 */
  voc = 'voc',
  /** DVD视频对象文件 */
  vob = 'vob',
  /** WAV音频文件 */
  wav = 'wav',
  /** Web应用程序归档文件 */
  war = 'war',
  /** WebM视频文件 */
  webm = 'webm',
  /** WebP图像文件 */
  webp = 'webp',
  /** Windows Media Audio文件 */
  wma = 'wma',
  /** Windows Media Video文件 */
  wmv = 'wmv',
  /** WavPack音频文件 */
  wv = 'wv',
  /** Excel电子表格文件 */
  xls = 'xls',
  /** Excel电子表格文件(XML) */
  xlsx = 'xlsx',
  /** XviD视频文件 */
  xvid = 'xvid',
  /** XZ压缩文件 */
  xz = 'xz',
  /** Z压缩文件 */
  z = 'z',
  /** ZIP压缩文件 */
  zip = 'zip',
  /** 扩展ZIP压缩文件 */
  zipx = 'zipx',
  /** Zstandard压缩文件 */
  zst = 'zst',
}